client/requirements.txt
Logbook==1.1.0
-e git://github.com/quantopian/MongoDBProxy.git#egg=mongodbproxy
psutil==5.2.2
pymongo==3.4.0
Pynetinfo==0.2.3
PyYAML==3.12
requests==2.18.1
